+++ apps/documentation/src/app/examples/toast/toast.example.ts
@@ -1,14 +1,14 @@
-import { Component } from '@angular/core';
-import { NgpToast } from 'ng-primitives/toast';
+import { Component, inject, TemplateRef } from '@angular/core';
+import { NgpToast, NgpToastManager } from 'ng-primitives/toast';
 
 @Component({
   selector: 'app-toast',
   imports: [NgpToast],
   template: `
-    <button class="toast-trigger" (click)="toast.show()" ngpButton>Show Toast</button>
+    <button class="toast-trigger" (click)="show(toast)" ngpButton>Show Toast</button>
 
-    <ng-template #toast="ngpToast" ngpToast let-dismiss="dismiss">
-      <div class="toast">
+    <ng-template #toast let-dismiss="dismiss">
+      <div ngpToast>
         <p class="toast-title">This is a toast message</p>
         <p class="toast-description">It will disappear in 3 seconds</p>
         <button class="toast-dismiss" (click)="dismiss()" ngpButton>Dismiss</button>
@@ -43,8 +43,17 @@ import { NgpToast } from 'ng-primitives/toast';
       background-color: var(--ngp-background-active);
     }
 
-    .toast {
-      position: fixed;
+    [ngpToast] {
+      position: absolute;
+      touch-action: none;
+      transition:
+        transform 0.4s,
+        opacity 0.4s,
+        height 0.4s,
+        box-shadow 0.2s;
+      box-sizing: border-box;
+      align-items: center;
+      gap: 6px;
       display: inline-grid;
       background: var(--ngp-background);
       box-shadow: var(--ngp-shadow);
@@ -53,11 +62,14 @@ import { NgpToast } from 'ng-primitives/toast';
       opacity: 0;
       transition: all 0.4s cubic-bezier(0.215, 0.61, 0.355, 1);
       border-radius: 8px;
-      z-index: 9999;
+      z-index: var(--ngp-toast-z-index);
       grid-template-columns: 1fr auto;
-      grid-template-rows: auto auto;
+      grid-template-rows: min-content min-content;
       column-gap: 12px;
       align-items: center;
+      width: var(--ngp-toast-width);
+      height: fit-content;
+      transform: var(--y);
     }
 
     .toast-title {
@@ -68,6 +80,7 @@ import { NgpToast } from 'ng-primitives/toast';
       grid-column: 1 / 2;
       grid-row: 1;
       line-height: 16px;
+      user-select: none;
     }
 
     .toast-description {
@@ -77,6 +90,7 @@ import { NgpToast } from 'ng-primitives/toast';
       grid-column: 1 / 2;
       grid-row: 2;
       line-height: 16px;
+      user-select: none;
     }
 
     .toast-dismiss {
@@ -93,33 +107,95 @@ import { NgpToast } from 'ng-primitives/toast';
       max-height: 27px;
     }
 
-    .toast[data-toast='visible'] {
-      opacity: 1;
-    }
-
-    .toast[data-position='end'] {
-      right: 16px;
-    }
-
-    .toast[data-position='start'] {
-      left: 16px;
-    }
-
-    .toast[data-gravity='top'] {
-      top: -150px;
-    }
-
-    .toast[data-gravity='bottom'] {
-      bottom: -150px;
-    }
-
-    .toast[data-position='center'] {
-      margin-left: auto;
-      margin-right: auto;
-      left: 0;
+    [ngpToast][data-position-x='end'] {
       right: 0;
-      max-width: fit-content;
+    }
+
+    [ngpToast][data-position-x='start'] {
+      left: 0;
+    }
+
+    [ngpToast][data-position-y='top'] {
+      top: 0;
+      --lift: 1;
+      --lift-amount: calc(var(--lift) * var(--ngp-toast-gap));
+      --y: translateY(-100%);
+    }
+
+    [ngpToast][data-position-y='bottom'] {
+      bottom: 0;
+      --lift: -1;
+      --lift-amount: calc(var(--lift) * var(--ngp-toast-gap));
+      --y: translateY(100%);
+    }
+
+    [ngpToast][data-enter] {
+      opacity: 1;
+      --y: translateY(0);
+    }
+
+    [ngpToast][data-exit] {
+      opacity: 0;
+      --y: translateY(calc(calc(var(--lift) * var(--ngp-toast-gap)) * -1));
+    }
+
+    [ngpToast][data-visible='false'] {
+      opacity: 0;
+      pointer-events: none;
+    }
+
+    [ngpToast][data-expanded='true']::after {
+      content: '';
+      position: absolute;
+      left: 0;
+      height: calc(var(--ngp-toast-gap) + 1px);
+      bottom: 100%;
+      width: 100%;
+    }
+
+    [ngpToast][data-expanded='false'][data-front='false'] {
+      --scale: var(--ngp-toasts-before) * 0.05 + 1;
+      --y: translateY(calc(var(--lift-amount) * var(--ngp-toasts-before)))
+        scale(calc(-1 * var(--scale)));
+      height: var(--ngp-toast-front-height);
+    }
+
+    [ngpToast][data-expanded='true'] {
+      --y: translateY(calc(var(--lift) * var(--ngp-toast-offset)));
+      height: var(--ngp-toast-height);
+    }
+
+    [ngpToast][data-swiping='true'] {
+      transform: var(--y) translateY(var(--ngp-toast-swipe-amount-y, 0))
+        translateX(var(--ngp-toast-swipe-amount-x, 0));
+      transition: none;
+    }
+
+    [ngpToast][data-swiping='true'][data-swipe-direction='left'] {
+      /* Fade out from -45px to -100px swipe */
+      opacity: calc(1 - clamp(0, ((-1 * var(--ngp-toast-swipe-x, 0px)) - 45) / 55, 1));
+    }
+
+    [ngpToast][data-swiping='true'][data-swipe-direction='right'] {
+      /* Fade out from 45px to 100px swipe */
+      opacity: calc(1 - clamp(0, (var(--ngp-toast-swipe-x, 0px) - 45) / 55, 1));
+    }
+
+    [ngpToast][data-swiping='true'][data-swipe-direction='top'] {
+      /* Fade out from -45px to -100px swipe */
+      opacity: calc(1 - clamp(0, ((-1 * var(--ngp-toast-swipe-y, 0px)) - 45) / 55, 1));
+    }
+
+    [ngpToast][data-swiping='true'][data-swipe-direction='bottom'] {
+      /* Fade out from 45px to 100px swipe */
+      opacity: calc(1 - clamp(0, (var(--ngp-toast-swipe-y, 0px) - 45) / 55, 1));
     }
   `,
 })
-export default class ToastExample {}
+export default class ToastExample {
+  private readonly toastManager = inject(NgpToastManager);
+
+  show(toast: TemplateRef<void>): void {
+    this.toastManager.show(toast, { placement: 'bottom-end' });
+  }
+}