+++ packages/ng-primitives/popover/src/popover-arrow/popover-arrow.ts
@@ -1,0 +1,12 @@
+import { Directive } from '@angular/core';
+import { setupOverlayArrow } from 'ng-primitives/portal';
+
+@Directive({
+  selector: '[ngpPopoverArrow]',
+  exportAs: 'ngpPopoverArrow',
+})
+export class NgpPopoverArrow {
+  constructor() {
+    setupOverlayArrow();
+  }
+}