+++ packages/ng-primitives/popover/src/popover/popover.ts
@@ -1,5 +1,6 @@
-import { Directive } from '@angular/core';
+import { Directive, input } from '@angular/core';
 import { NgpFocusTrap } from 'ng-primitives/focus-trap';
+import { explicitEffect } from 'ng-primitives/internal';
 import { injectOverlay } from 'ng-primitives/portal';
 
 /**
@@ -11,10 +12,12 @@ import { injectOverlay } from 'ng-primitives/portal';
   hostDirectives: [NgpFocusTrap],
   host: {
     role: 'dialog',
+    '[id]': 'id()',
     '[style.left.px]': 'overlay.position().x',
     '[style.top.px]': 'overlay.position().y',
     '[style.--ngp-popover-trigger-width.px]': 'overlay.triggerWidth()',
     '[style.--ngp-popover-transform-origin]': 'overlay.transformOrigin()',
+    '[attr.data-placement]': 'overlay.finalPlacement()',
   },
 })
 export class NgpPopover {
@@ -22,4 +25,13 @@ export class NgpPopover {
    * Access the overlay.
    */
   protected readonly overlay = injectOverlay();
+
+  /**
+   * The unique id of the tooltip.
+   */
+  readonly id = input(this.overlay.id());
+
+  constructor() {
+    explicitEffect([this.id], ([id]) => this.overlay.id.set(id));
+  }
 }