+++ packages/ng-primitives/utils/src/helpers/attributes.ts
@@ -1,0 +1,16 @@
+import { afterRenderEffect, Signal } from '@angular/core';
+
+export function booleanAttributeBinding(
+  element: HTMLElement,
+  attribute: string,
+  value: Signal<boolean> | undefined,
+): void {
+  if (!value) {
+    return;
+  }
+
+  afterRenderEffect({
+    write: () =>
+      value() ? element.setAttribute(attribute, '') : element.removeAttribute(attribute),
+  });
+}